Add files via upload
Inthis concept of Polynomial non-linear regression we studied that A regression model where the relationship between X and Y is modeled as an nth degree polynomial.
where, to get the curved or polynomial from it in non-linear format. Then,fit the model to convert into non-linear to linear regressio form.