Add files via upload
In this algorithm of decision tree regression learnt about Splits the dataset into branches at decision nodes based on feature values and predicts by averaging
the outputs in a leaf.
Also ti is in non-linear form. It has process of Entrophy and information gain and also it have process of  purning to remove un-necessary part from it.